move interfaces.d.ts
