+++ .gitignore
@@ -1,5 +1,6 @@
 .git
 node_modules
-*.js
 .vscode
 jsOut
+test*.js
+*.js
