+++ .npmignore
@@ -4,7 +4,5 @@
 docs
 node_modules
 .eslintrc.json
-tsconfig.json
-.gitignore
 test*.js
-*.ts
+*.js
