+++ src/races/halfelf.ts
@@ -1,5 +1,5 @@
 
-import { Stats } from "../functions/interfaces.js";
+import { Stats } from "../../interfaces.js";
 
 
 export default async () => {
